backend/requirements.txt
fastapi>=0.115
uvicorn>=0.34
mysql-connector-python>=9.0
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
python-dotenv>=1.0
openai>=1.0
pandas>=2.0
numpy>=2.0
openpyxl>=3.1
